Solution to Determine Eligibility for Voting

+++ Determine Eligibility for Voting@@6788dd63490637cfc44e1b99/code.c
@@ -5,7 +5,7 @@ char* welcome() {
 }
 
 int main() {
-    float a;
+    int a;
     scanf("%d",&a);
     a>=18?printf("Eligible"):printf("Not Eligible");
 }
